requirements.txt
Flask
Flask-Cors
numpy
pandas
scikit-learn
networkx
nltk
wordcloud
matplotlib
seaborn
statsmodels
Pillow